Cleaning up empty code, moving only SOLEIL readers first
